[test] Separate unit tests by API entry points
